+++ docs/internationalization.rst
@@ -1,3 +1,7 @@
+.. automodule:: xclim.core.locales
+   :members:
+   :show-inheritance:
+
 .. automodule:: xclim.locales
    :members:
    :show-inheritance:
