+++ xclim/core/units.py
@@ -19,8 +19,8 @@ import xarray as xr
 from boltons.funcutils import wraps
 from packaging import version
 
-from xclim.core.options import datacheck
-from xclim.core.utils import ValidationError
+from .options import datacheck
+from .utils import ValidationError
 
 
 __all__ = [
