+++ xclim/indices/_simple.py
@@ -1,10 +1,10 @@
 import xarray
 
+from . import run_length as rl
 from xclim.core.units import convert_units_to
 from xclim.core.units import declare_units
 from xclim.core.units import pint_multiply
 from xclim.core.units import units
-from xclim.indices import run_length as rl
 
 xarray.set_options(enable_cftimeindex=True)  # Set xarray to use cftimeindex
