+++ xclim/indices/_threshold.py
@@ -3,12 +3,12 @@ import datetime
 import numpy as np
 import xarray
 
+from . import run_length as rl
+from .generic import threshold_count
 from xclim.core.units import convert_units_to
 from xclim.core.units import declare_units
 from xclim.core.units import pint_multiply
 from xclim.core.units import units
-from xclim.indices import run_length as rl
-from xclim.indices.generic import threshold_count
 
 xarray.set_options(enable_cftimeindex=True)  # Set xarray to use cftimeindex
